okmovieSpider/requirements.txt
beautifulsoup4==4.6.3
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
Flask==1.0.2
geoip2==2.9.0
get==1.0.3
getproxy==0.2.3
gevent==1.3.7
greenlet==0.4.15
idna==2.7
itsdangerous==0.24
Jinja2==2.10
lxml==4.2.5
MarkupSafe==1.0
maxminddb==1.4.1
mysql-connector-python==8.0.13
post==1.0.2
protobuf==3.6.1
psutil==5.4.7
public==1.0.3
pycparser==2.19
query-string==1.0.2
request==1.0.2
requests==2.19.1
retrying==1.3.3
six==1.11.0
urllib3==1.23
virtualenv==16.0.0
Werkzeug==0.14.1
